fix: address Copilot review — guard refetch in waiting state, floor duration minutes, float db_max_size_mb, fix test comment

+++ src/hassette/config/config.py
@@ -280,7 +280,7 @@ class HassetteConfig(ExcludeExtrasMixin, BaseSettings):
     db_retention_days: int = Field(default=7, ge=1)
     """Number of days to retain execution records (handler_invocations, job_executions)."""
 
-    db_max_size_mb: int = Field(default=500, ge=0)
+    db_max_size_mb: float = Field(default=500, ge=0)
     """Maximum database file size in MB. When exceeded, oldest execution records are deleted. 0 disables the size
     failsafe."""
